[pdpix] Bug Fix: Don't Optimize Size of Structures
This commit fixes
https://github.com/demikernel/demikernel/issues/704

+++ examples/tcp-echo/server.rs
@@ -169,10 +169,10 @@ impl TcpEchoServer {
     fn handle_fail(&mut self, qr: &demi_qresult_t) -> Result<()> {
         let qd: QDesc = qr.qr_qd.into();
         let qt: QToken = qr.qr_qt.into();
-        let errno: i32 = qr.qr_ret;
+        let errno: i64 = qr.qr_ret;
 
         // Check if client has reset the connection.
-        if errno == libc::ECONNRESET {
+        if errno == libc::ECONNRESET as i64 {
             println!("INFO: client reset connection (qd={:?})", qd);
             self.handle_close(qd)?;
         } else {
