+++ examples/tcp-wait/client.rs
@@ -84,7 +84,7 @@ impl TcpClient {
             // Wait for push().
             match self.libos.wait(push_qt, None) {
                 Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_PUSH && qr.qr_ret == 0 => {},
-                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED => {},
+                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED as i64 => {},
                 _ => anyhow::bail!("wait() should succeed with push() after async_close()"),
             }
         }
@@ -109,7 +109,7 @@ impl TcpClient {
             // Wait for pop().
             match self.libos.wait(pop_qt, None) {
                 Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_POP && qr.qr_ret == 0 => {},
-                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED => {},
+                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED as i64 => {},
                 _ => anyhow::bail!("wait() should not succeed with pop() after close()"),
             }
         }
@@ -133,7 +133,7 @@ impl TcpClient {
             // Wait for push().
             match self.libos.wait(push_qt, None) {
                 Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_PUSH && qr.qr_ret == 0 => {},
-                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED => {},
+                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED as i64 => {},
                 _ => anyhow::bail!("wait() should not succeed with push() after close()"),
             }
         }
@@ -158,7 +158,7 @@ impl TcpClient {
             // Wait for pop().
             match self.libos.wait(pop_qt, None) {
                 Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_POP && qr.qr_ret == 0 => {},
-                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED => {},
+                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED as i64 => {},
                 _ => anyhow::bail!("wait() should not succeed with pop() after close()"),
             }
         }
@@ -176,7 +176,7 @@ impl TcpClient {
             // Wait for push().
             match self.libos.wait(push_qt, None) {
                 Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_PUSH && qr.qr_ret == 0 => {},
-                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED => {},
+                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED as i64 => {},
                 _ => anyhow::bail!("wait() should succeed with push() after issuing async_close()"),
             }
 
@@ -203,7 +203,7 @@ impl TcpClient {
             // Wait for pop().
             match self.libos.wait(pop_qt, None) {
                 Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_POP && qr.qr_ret == 0 => {},
-                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED => {},
+                Ok(qr) if qr.qr_opcode == demi_opcode_t::DEMI_OPC_FAILED && qr.qr_ret == libc::ECANCELED as i64 => {},
                 _ => anyhow::bail!("wait() should not succeed with pop() after close()"),
             }
